requirements.txt
astroid==3.2.4
beautifulsoup4==4.12.3
certifi==2024.7.4
cfgv==3.4.0
charset-normalizer==3.3.2
dill==0.3.8
distlib==0.3.8
filelock==3.15.4
identify==2.6.0
idna==3.7
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
mysql-connector-python==9.0.0
nodeenv==1.9.1
packaging==24.1
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.8.0
pylint==3.2.6
pytest==8.3.2
PyYAML==6.0.2
requests==2.32.3
requests-mock==1.12.1
soupsieve==2.5
tomlkit==0.13.0
urllib3==2.2.2
virtualenv==20.26.3
